First Commit
Do not finish level 3, the game crashes as I never defined level_four properly and as a result it must be force quit

Building from what I learned from Missy Munch, I wanted to make a platformer that features my other cat "Peeps". This was a lot harder and most of the information I was looking for was not in the youtube videos I was watching as guides, which led to me downloading a few other platformer games on Github and testing some of their features.

The plan was to start with my basic knowledge on level 1, and then improve the game every level - however the problem with this as I learned by level 3 that by storing all of the levels within one file, it becomes a nightmare to find your way around. Also rules that you made for the first level will impact future levels unless defined every time, in future if i tried this again I would separate levels into different files, so any functions could be more specified.

artwork made in Piskel, music (if i added it) is probably just copied from the "Missy Munch" game i previously made